requirements.txt
connexion[swagger-ui] >= 2.6.0
Flask >= 2.0.0
Flask-SQLAlchemy >= 2.5.0
Flask-Login >= 0.5.0
Flask-Bcrypt >= 0.7.1
PyMySQL >= 1.0.2
swagger-ui-bundle >= 0.0.2
python_dateutil >= 2.6.0
setuptools >= 21.0.0
WTForms >= 2.3.3
PyJWT >= 2.1.0
Flask-CORS >= 3.0.10
Flask-Limiter >= 1.4
python-magic >= 0.4.24
bleach >= 4.1.0
